Move feature toggles out of the dungeon to commons package
Feature toggles are commonly used across the modules and there is no
reason to keep them in a package that is not related to what they are
doing.
