+++ community/common/src/main/java/org/neo4j/util/FeatureToggles.java
@@ -17,7 +17,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.unsafe.impl.internal.dragons;
+package org.neo4j.util;
 
 import java.util.Objects;
